Winkelgebieden: translation of generated PK col name to 'ID'
To meet the prerequisite of the Amsterdams Schema of having an 'ID' attribute, the creation of the winkgeb table is adjusted so the generated PK column 'ogc fid' is translated to 'id'
The translation is realized by adding the -lco (Layer creation option) option to the ogr2ogr command. The -lco takes an attribute identifier as input and its transalation. I.e. FID=ID.