winkelgebieden: changed dag_id to "winkelgebieden" and folder data/winkgeb to "data/winkelgebieden"
To comply with the naming convention, the database table for winkelgebieden is named "winkelgebieden_winkelgebieden". This also means that the DSO api endpoint for winkelgebieden will be /winkelgebieden/winkelgebieden.

winkelgebieden: changed dag_id to "winkelgebieden" and folder data/winkgeb to "data/winkelgebieden"

To comply with the naming convention, the database table for winkelgebieden is named "winkelgebieden_winkelgebieden". This also means that the DSO api endpoint for winkelgebieden will be /winkelgebieden/winkelgebieden.