[GREX] changed startime and convert geometrycollections
- Starttime is change from default to 0600 a.m.
- Source contains geometrycollections (polygons and points in one object)

The startime has changed to a line with the data processing schedule of
DWH stadsdelen. Which provide us with GREX data.

Due to geometrycollections (probably due to source errors) some records could not be processed correctly.
The effect was that these records where not processed.
In order to show the underlining polygons, the additional points in the object are removed.
The order of the convertion steps were not correctly defined.