[DS-563] add dag_id => dataschema name translation
Some DAG's do not match dag_id with dataschema name.
The reason can be that for one dataset multiple DAG's are implemented i.e. for each table one DAG.
Since the dataschema is the basis for the database permissions, for the situation where dag_id does not match dataschema name, the relation between the dag_id and schema name is held in this constant.