Install wheel before other pip installs [DS-559]
When the wheel package is installed, subsequent pip installs of other
packages will build wheels if they are not yet available and caches
them. This is especially useful on platforms that generally don't have
many prebuilt wheels made available for them.

Should it be necessary to recreate the virtual environment during
development these cached wheels can to be used to install the packages
instead of rebuilding them one by one.