Add XCom support to HTTP Fetch Operator. [DS-559]
The tmp dir can now be communcated to the HTTP Fetch Operator by means
of XCom. This is necessary if we want to be able to dynamically create
temporary directories (in other operators), instead of relying upon hard
coded temporary directories. The old way is still supported though.

Also rearranged the code so that the _precariobelasting_ specific code
is now a single block. Should make it easier to rip it out when this
workaround is not needed anymore.

Lastly documented the two different ways of passing in a temporary
directory/file. Didn't bother documenting existing undocumented
functionality.