Added PostgresCreateTablesLikeOperator [DS-559]
This operator allows one to create new tables based on the definition of
existing tables. Specifying the set of existing tables is done by means
of a regex. The operator uses proper mechanisms to construct SQL queries
(parameter *and* identifier escaping). Whereas many other operators use
basic string formatting which is not safe (SQL injection).

The operator is also resilient in the presence of existing tables that
it intended to create. Instead of failing, it will capture the notices
issued by PostgreSQL and logs them.