Start with a clean sheet by dropping tmp table in advance. [DS-559]
If a previously DAG run failed and didn't get around to cleaning up the
temporary tables, the next run will fail as well. This, due to trying to
insert values in the tmp tables that are likely already there and
violates uniqueness constraints. Hence before we do anything else that's
DB related, we clean up what might (hence the `IF EXISTS`) be left
around.