Reworked BBGA DAG to support `pg_comparator`'s `--pg-copy`. [DS-559]
BBGA's largest table has over 6 million rows in it. The primary key of
that table is of type `VARCHAR`. As it turns out `pg_comparator` is not
able to use its much faster `--pg-copy` method on keys (primary or
alternate) of that type. It requires them to be of type integer.

To solve that problem we create yet another column named `cdc_id` of the
largest integer type supported by `pg_comparator`. In it we store an 8
byte hash (using BLAKE2s) of the original primary column. Next we invoke
the reworked `PgComparatorCDCOperator` to use that column with the
`use_pg_copy` option.

The performance improvement is massive. In an Bhyve virtual machine
running Alphine Linux, running the database container, the original
PgComparatorCDCOperator initial run took 1hr26min. With the above
optimizations the initial run only takes 6mins. That's 14x faster!!!

Thanks to @cvriel for pointing out the existance of the `--pg-copy`
option.