Additional pre-commit checks, add mypy config & dependency. [DS-559]
Enabled some additional pre-commit checks that are low impact.

Although type annotations have proven themselves to be a massive help
with regards to:

- typing in code (better IDE suggestions)
- creating more robust code
- communicating to other developers what code expects and returns

It seems to be used rather haphazardly in this project. Before turning
on a `pre-commit` hook that enforces mypy type checks, we should first
work with `mypy` by adding a suitable configuration and the means to run
it from the commend line.

Mypy can now be run as follows:

    $ mypy src/dag/bbga.py

Or with brace expansions to let it loose some on the files we've worked
on:

    $ mypy src/dags/bbga.py \
src/plugins/{postgres_{create_tables_like_operator.py,\
insert_csv_operator.py},sqlalchemy_create_object_operator.py,\
http_fetch_operator.py}