Do enable mypy in pre-commit.
It's too easy to make a mess of type annotations if we don't gradually
enforce them. By adding it to the pre-commit hook anything new should be
properly type annotated. This should slowly but steadily migrate the
code base to an annotated one.

BTW it's a shame Airflow is so lightly annotated.