Per @ngaranko: don't use per DAG run specific tmp dirs.  [DS-559]
Instead use tmp directory shared by all runs of a specific DAG.

Eg, all runs of the BBGA DAG will now use: `/tmp/bbga` instead of each
run using `/tmp/bbga_<random string>`. This to prevent tmp directories
being left behind if DAG runs are stopped or fail half way through.

Although currently we use `/tmp` voor the root of the tmp directories,
for which the OS has the means to clean it up (a reboot generally
suffices) this will likely change in the future. And as we deal with
potentially large datasets stored in these tmp directories we could end
up wasting storage that is not readily reclaimed.