Remove `ind_drop` from `PostgresTableCopyOperator`.
The `ind_drop` parameter was not use anywhere, and if it was would have
broken code. Reason being that the conditional expression that
evaluated `ind_drop` would have resulted in a `None` value. Later on
`.format` would have been called on that `None` value resulting in an
`AttributeError`.

As the feature wasn't used I decided to remove it instead of
fixing it.