Rework `PostgresTableCopyOperator`. [DS-559].
This operator does many things. For it to be more useful to other DAGs
I made the following things optional:

- truncating target table
- copying data
- dropping source table

This way it can be used to duplicate table definitions only, as is
required for the BBGA DAG

In addition removed the `check_safe_name` checks. These are
unnecessary when using the `psycopg2.sql` library.

Also added logging so that it is clear what this operator did from the
logs.

By not using `PostgresOperator` we can easily control transaction
management; ensure all operations run in the same transaction.