Use `PostgresTableCopyOperator` to duplicate table defs. [DS-559]
Now that `PostgresTableCopyOperator` has been reworked to make most of the
things it does optional, we can use it for the simple task we require it
to do: copy table definitions.

Contrary to `PostgresCreateTablesLikeOperator` that could duplicate many
source tables at once, `PostgresTableCopyOperator` only operates on one
source table (+ its junction tables). Hence we need to create as many
tasks for this operator as we have table to duplicate. However as the
previous set of tasks were parallel set as well we needed a
`DummyOperator` to connect two sets of parallel tasks.