Accomodate the use of shortnames in schemas
In the previous implementation, the dataset was only needed in the
HttpGobOperator. The names of tables involved in the import process
were having fixed names and were derived from the names of the
graphql definition files.

In the new situation, table (and column) names can be overridden from the
amsterdam schema. So, the dataset schema information is needed
throughout the whole DAG (e.g. for table renaming).

It is bad practice to load the schema information directly in the
dag module, beause this module is reloaded numerous times by
Airflow. All processing should be done exclusively in operators.

So, a new operator had been added at the start of the DAG
that reads the relevant schema and put this information on XCom.

Operator later in the DAG that need information about table names
can fetch this information from XCom.

This involved the modification of several operator more
downstream in the DAG. Those operator were relying on table
names entering the operator during construction.
However, when a `shortname` is used in the schema definition,
the tablename needs to be obtained from XCom.

So the relevant operators have been modified to accommodate this
behaviour.

To keep the code as DRY as possible, this XCom-related behaviour
has been put in a mixin class.

To keep the code generic, a function can be injected into these
operators that is able to modify the relevant operator attributes
(usually attributes holding table names).