Pipline deelmobiliteit
So far 3 resources are added to dataset deelmobiliteit. The resources are categorized into: cars, scooters (and bikes(still to come)). More resources to come. ~ Eleven in total.

Data classes are used to structure the data. Geometry objects are transformed to the appropiate geometry datatype. Source is processed from an API and imported in the database.

The data is stored as slowly changing dimensions (sdc-2). Each call leads to new records. The time frame data is kept, is set to one month. Therefore the CDC operator doesn't process deletes only inserts and updates.