changed http_fetch_operator for maps.amsterdam.nl and sport update
Currently one of the data sources of sport is failing. To cope with the situation
the http_fetch_operator is changed so it can handle the maps.amsterdam.nl source.

Also in the sports datasets an extra step for cleaning up the data is added (on request of the user).