Pipeline RDW
The RDW source contains of - in this particular case - of 3 resources. Each of those resources contains about 15 million records. To make the data processing a fast as possible the data is downloaded as a comma seperated file (csv). This is by far the most efficient approach. Alternatives where several API calls returning the data in batches and commiting them as a batch to the database (consuming a lot of memory). However this took way too much time.