[WIOR] added an upload step to store the data in the objectstore
Currently the WIOR (werken in openbare ruimte) data is partially served and processed in the database. To store the complete 'raw' data, its uploaded after download into the objectstore. The files are flagged with the current day so a history of data processing can be registred and serviced if needed to other consumers.