AB#4367 Ditch unnecessary use of formatted string literals.
Formatted string literals should be used when a string needs to be
formatted; as in, composed from a string interspersed with variable
values. If it it just the variable that already happens to be of type
string, a formatted string literal has no use. It conveys the wrong
intention.

Nor should they be used (though not the case here) to convert
non-strings into string. Use `str()` for that as that conveys the
intention better.