Schema versions not accounted for in the create table operator
The first versioned schema 'huishoudelijkafval' is in operation. In the Airflow dag the schema is used to create the database table. But the operator is based on the assumption that all schema's are located in the path [schema_name]/[schema_name]. Due to versioning that is no longer the case. There are subfolders containing the versionumber as the name i.e. [schema_name]/[version_number]/[schema_name]. The operator now looks explicit to the index.json file in the schema objectstore the lookup the path of the schema. If needed a specific version number can be added the setup the path.