AB#6837 table creation solely based on schema
The existing DAG beheerkaartbasis was lacking the table creation based on the schema defintion. Also the whishes of the users were to create the database objects and it's comments on the schema defintion structure and it's description fields.

The existing DAG was enriched by the extra step of creating the database objects. However, the source data are SQL copy statements that reference a different schema then public. So the tables to be created had to be created in the specified schema as the source copy statements dictated (the PTE schema). The schema option was already available as a paramater in the create objects method from schema-tools. So it could be set.

Also temorary database tables names, as stated in the copy SQL statements from source file, was an option to set. This way the data structure could be created solely from the schema and the source SQL copy statements could be excuted to insert the data.