AB#4367 Clean pre-commit env.
The `pre-commit` `autoupgrade` command, will upgrade the hooks to their
newest versions. However it will not touch the
`additional_dependencies`. This is an issue for `flake8` that depends on
a large number of them. We address that by occasionnally cleaning the
`pre-commit` environments (that where `pre-commit` installs the hooks +
plugins). The `install` Makefile target seems like a good location for
that. Afterall we run it after having upgraded project dependencies.

Downside is that after having run `make -C src install` the next `git
commit` will have to recreate the `pre-commit` environment which will
take a bit of time.