Moved `set -e` closer to execution scripts/checkdags.py
The set -e configuration in run.sh is used to stop Airflow when one of the DAG's cannot be executed due to some error. An extra safety check before deployment of changes.

In the run.sh one of the cmd's is to delete and create Airflow connections. When runned locally (for the first time) the connection are not yet created when the cmd to delete the connection is executed (after that it will create the connection).

Because the connection does not yet exists when deleted, it will give an exit code 1. Causing the Airflow to stop running.

As a quick fic the `set -e` is set just befor e the scripts/checkdags.py. So it will be active for the check on correct DAG.