AB#8701 Add new mk_dir @task-based task.
This task uses the new Airflow Task API. In addition it uses plain
stdlib functions to create, and optionally clean, a directory. This is
an improvement over how we currently created (temporary) directories in
Airflow: spawning a shell process only to invoke the `mkdir` commandline
tool.

Unfortunately the `@task` decorator in untyped. Hence the need for a
type ignore comment.