AB#8701 CREATE TABLE with LIKE INCLUDING ALL.
iNCLUDING ALL, as we want DEFAULT values, COMMENTS and IDENTITY as well.
DEFAULT values and IDENTITY are essential to provide meaningfull and
easy to manage auto incrementing values as used by the `anpr` DAG. And
COMMENTS we should just keep. After all, if `schema-tools` bothers to
generate them for us, we should at least try to keep them.

We do get `STORAGE` along with using `ALL`. Not sure what the impact of
that is. However should that be a problem, we can alway use something
along the lines of:

  INCLUDING ALL EXCLUDING STORAGE

or whatever option you want excluded.