AB#8701 Schema based table creation for anpr.
The `anpr` DAG now creates its table based on its own schema by using
`SqlAlchemyCreateObjectOperator` instead of relying on an external
system (DSO-API) to do so. In addition it now uses
`PostgresTableCopyOperator` instead of relying on custom SQL.

The usage of a sequence to automatically generated values for the `id`
column proved cumbersome in the presence of
`SqlAlchemyCreateObjectOperator` and `PostgresTableCopyOperator`.
Fortunately, from version 10 on, PostgreSQL has identity columns that
work much nicer and are easier dropped and copied; features that this
DAG depends on.