Rename identically named non-package modules in src tree
In the source tree we have a `vsd` directory. Although it lives under
`src` it is not a proper Python package. This cause import problems
when, somewhere below the `vsd` directory, there are modules that are
identically named.

Though we don't regularly import things from the `vsd` directory, MyPy
does when you want to run it over the entire project with `mypy src`. To
allow MyPy to do its thing we have renamend these identically named
modules.