AB#8704 Schema based table creation for cmsa.
The `csma` DAG now creates its table based on its own schema by using
`SqlAlchemyCreateObjectOperator` instead of literal SQL statements in a
separate SQL file.

The former literal SQL statements created foreign keys between the
tables. We have chosen not to replicatie that in this rework of the DAG.
Reason being that although it migt be applicable for this particular
DAG, it is not something we can enforce with our schema based approach. As
such it would weird enforce it outside of the schema.

In addition:

- dropped the PorvenanceRenameOperator as the schema based creation does
  this automatically
- Switched to using identity based columns instead of serial based ones
- Inlined the SQL for the `cmsa_markering` table INSERT statement so
  that we could parameterize it.