OGR2OGR promote_to_multi fix
The promote_to_multi option was not optional yet. Turn it into an boolean value where de default is set to False (not promoting geometry objects to multigeometry objects).

Currently it was leading to errors in Airflow because it was defaulted to always promote. This lead to geometry check issues.