Hotfix: disable checkdags script.
Service does not come up after new deployment to acceptance. The
checkdags script might be a contributing factor to this issue as it
takes a while to run. This prevents the service to respond to the
deployment health checks in time.

But the real culprit is probably `run.sh` in its entirety. It takes way
too long to execute all those `airflow` commands.