AB#8706 Simplify Ogr2OgrOperator SQL usage.
Previously the operator required the SQL statement passed to it to be
quoted. But as Python strings are quoted as well this resulted in having
to escape the quotes around the SQL statement. By doing the quoting of
the SQL statement inside the operator, we don't have to do it at the
call site anymore. This simplifies its usage somewhat.