AB#8706 Don't use string formatting for SQL statements.
We really, really should not be using string formatting for SQL
statements! The libraries we use have ample facilities for doing it
properly: relying on database specific libraries and on the database
itself to do proper quoting and escaping. Using these libraries we guard
ourselves against SQL injection.

One might argue that the chance of actually having to deal with SQL
injection is rather low as we are not really dealing with direct user
input here. But limiting potential attack vectors is generally a good
thing. And as code tends to get copied around we should always craft it
properly to start with.