AB#8702 Call format on `SQL` not on `str`
Commit 0d2a486653afefa07d595f7764e442e4c43f64a7 provided a fix for the
issue where `sql.Identifier`s were inserted with their string
representation in the generated SQL statements. Eg:

    ALTER TABLE Identifier('public',...

Although the fix provided in the above commit solved the problem. The
actual problem was caused by inadvertently calling `format` on the `str`
instead of on the `SQL` object. Eg:

* Wrong:   `sql.SQL("...".format(...))`

instead of:

* Correct: `sql.SQL("...").format(...)`

This commit removes the `as_string` fix/workaround in favor of calling
`format` on the `SQL` object.