bouwstroompunten fix
The bouwstroompunten uses a HttpFetchOperator to call an API to retrieve source data.For using the API a API token is needed. This token was added to the header. By adding a get_token() within the header definition.

However during 'compile' (when importing the dags) the get_token() was evaluated. Resulting an a error message because the parameters used are not yet loaded.

To prevent from evaluting at compile time (instead of only at run time) the output of get_token() is now passed as a xcom variable. Based on the same logic as BBGA for consistency.