AB#8702 Clean up LogMessageOperator.
The `__init__` method had many unused parameters. Turns out they were
only there to keep them from ending up in `kwargs`. This is error prone
as it means this  `__init__` needs to be kept in sync with that of the
`SlackWebhookOperator`. Better to query `BaseOperator`'s `__init__`
method to see what it accepts and only pass that along.