AB#8702 Rewrite basiskaart to be one DAG with parallel tasks.
In preparation for schema based table creation I have changed the
dynamic DAG creation (a DAG per table) to be a regular single statically
defined DAG where all table operations are now parallel tasks. Reason
being that the `SqlAlchemyCreateObjectOperator` creates all tables in
the schema at once. Hence only needs to be called once and not once in
every dynamically created DAG.

NOTE: the `SqlAlchemyCreateObjectOperator` is not yet used in this
commit. This commit is just pre-work.