AB#8702 Do specify column names for DB inserts.
Schema based DB table creation creates the columns in a different order
than the previous literal SQL table create statements. This made the DB
inserts fail. By specifying the names of the columns in the correct
order upon insert, the order of the columns in actual table does not
really matter anymore.

The Zen of Python states:

    "Explicit is better than implicit."

Which in this particular case translates to:

    "Do not rely on the implicit order of columns. It will bite you!"