Add `drop_target_if_unequal` flag to PostgresTableCopyOperator
When a schema changes, the table copy statement

    INSERT INTO <target-table> SELECT * FROM <source-table>

will fail, because the number of columns does not match anymore.

This can be fixed by dropping the `target-table` altogether,
however, this could lead to missing data in production for much too
long.

Alternatively, the `target-table` can be fixed manually using an `ALTER`
to fix the column differences.

Option 1 is not acceptable. Option 2 is error-prone
and a lot of repetitive work.

So, an extra flag has been added to the table copy operator.
When this flag is set, the `source-table` and `target-table`
are compared before the data copying starts.
If the columns do not match, the `target-table` is dropped.
This automatically leads to creation of a new table by the
`INSERT INTO` statement.

And, because this all happens during an atomic operation,
there is no downtime for the delivery of data.

Review suggestions integrated:

Add better comparison for two tables

As suggested by @lbam.
And added check for the existence of the target table.

Simplify SQL statements as suggested by @gkoller
Only use `sql.SQL` if appropriate.

Furthermore: Bump schematools to 0.22.1