AB#8703 Update to API breaking schema-tools version 0.23.2
Version 0.23.2 of `schema-tools` broke the `schematools.utils` API
without reflecting that in its semver version number. As a result a
simple bump of the `schematools` dependency required adjusting a number
of calls into this package.

As there were no functional changes I didn't bother fixing all flake8
and mypy errors in the files that required thir `schematools.util` calls
to be fixed.