Transform BBGA CSV files using SQLite.
Reading in CSV files, transforming the data contained in them and
writing it out to new CSV files, all in Python, is rather slow.
Fortunately SQLite can do it too and much, much faster (seconds versus
serveral minutes). And the code to do so is rather succinct.

This Python module contains a multiline string consiting of all SQLite
commands and SQL queries to perform the transformation.