Use SQLite in BBGA for CSV transformations + table renames.
With the removal of the CDC Operator, filling the `bbga_kerncijfers`
table with 6+ million rows had become an extremely slow process as one
was using the table copy operator. By using the table rename operator
things have become fast again..

Another optimization is the usage of SQLite to:
- read in CSV files
- perform transformations on the tables (aka compensating for Django's
  lack of composite primary keys)
- writing the resuling transformations out as CSV file again.

This is much, much faster than doing that in pure Python.