Add sqlite3 > 3.33 to docker image
It turned out that the docking image was missing the sqlite3 binary.
However, adding it using `apt install` still did not solve the problem.

That is because the "generated columns" feature that is being used
is introduced in sqlite 3.33. The base of the Docker image that is being used
in Airflow has a sqlite3 binary that is too old (32.27.2).

So, a newer sqlite3 (version 3.36) is being compiled from source during
construction of the Docker image.

Implemented review comments LB:

- More efficient config/make/make install
- Add sha check on downloaded sqlite3 src tarball