[FIX] postgres_table_copy_operator - column order
Some DAG''s crashed due to a mismathed datatype when using the postgres_table_copy_operator. The reason is that the operator made the assumption that the column order of the target and the source table during a INSERT SQL statment are identical. However this does not is alwyays the case now and maybe in the (near) future also. To prevent a mismatch error the columns are added into the INSERT explictly so data can get inserted with the correct mapping despite column order.