swiftoperator explicit connection to VSD objectstore
In some DAG''s the swiftoperator is used to download data from the VSD objectstore. The parameter `swift_conn_id` is in many cases omitted. Because with the parameter it defaults to the VSD objectstore. However this default swift (objectstore) connection is not using the baseoperator.get_connection(). This has a drawback on using KeyVault on the Airflow instance on AKS. Basicly KeyVault is not consulted. When ommitted it will use the swift API that will look for OS_* env vars. The Airflow on AKS based on a HELM installation is expecting the Airflow naming conventions for env vars. Hence, that would be AIRFLOW_CONN_*

To solve this, every DAG should always explicitly add swift_id as the `swift_conn_id`. Then the Airflow native baseoperator.get_connection() will be used.