Fix bug in PostgresTableCopyOperator
The table copy operator copies the content of a source table
to a target table. A lot of data ingestion is done by first
creating and filling a temporary table and then copy-ing over
the data to the final destination, the target table.

When there is a schema change, this fails. There is a special flag
that drops the target table in case of differences in the column
structure between source and target.

However, also a change in column order was triggering a drop
of the target table.

And enhancement was added by Chris, that determines the
structure of the target table from the amsterdam schema.
So, as long as the set of columns in source and target
table is equals, no dropping of the target is needed.

But, this change did not take into account the junction tables
that are created for relations between dataset tables.

This bugfix is fixing that omission.