from typing_extensions import Required
Oracle DB connection

Some of the resources use a Oracle DB that is hosted by KPN.
Currently these database require a specification of a SID of SERVICE NAME to connect. Containerized DB servers require SERVICE NAME.
The latter was not yet processable in current logic. To make sure that we can continue the processing of data in an Oracle database (that require now a service name), adjustments where made.