Oracle source database connection
Due to the migration of KPN Oracle database's and the need to connect to those database with a service_name instead of a SID, the get_ora_engine() method in the common/db.py was modified. It now uses the default (of the shelf) AirflowHook, which handles the connection with the given service_name. There still the possibility to use the SID (or DB name after the port number) as a valid connection string to Oracle databases that can be accessed by a SID or DB name.

In short:

- get_ora_engine() method uses now the default OracleHook. Handeling a database connection.
- Moved the wkt_loads_wrapped() to db.py as a common used method. Reduces duplication in code and making it generic(-ish)
- Serval Black, Flake8 and mypy modifications to DAGS grex and horeca_exploitatievergunning while we add it...