beheerkaart not using correct database objects names
Since the change in Amsterdam schema definition of beheerkaart, the dataset uses a camelCase as in `beheerkaartBasis`. The DAG that utilizes operators that for instance create a database object needs a snake cased equivalent. Also due to creating of objects in a database schema other then `public` as in `pte` the creating of tables in the explicit correct database schema was not in place leading to errors.