Add a rename column for the PK of the woningbouwplan_buurten
The PKs of through tables should be called `id` due to a recent
change in DSO-API code. This commit does a column rename in the
database table that is loaded from the dump.

The owner of the dataset has been asked to rename this column
in the source (the incoming dump). If that is done,
the ALTER statement needs to be removed.
(PostgreSQL does not have a RENAME IF EXISTS).