Modified Asbestdaken to new name
The DAG asbest was refering to asbestdaken. The latter being a recent change on the Amsterdam schema definition. Because the dataset was renamed from `asbest` to `asbestdaken` the permissions operator failed because it uses the DAG name for looking up the dataset name. So the DAG name must match the dataset name. If not we can specify a rename in the permissions operator. But that would be a last resort.