Refactor scheduling for GOB dag
The GOB dag is using an operator that triggers the next
DAG as a final step in the DAG run. This makes all the dynamically
generated GOB dags run in some kind of "estafette" fashion,
strictly serializing the DAGs. The reason for this is that
the GOB API cannot handle much load.

However, not all datasets need to run according to the same schedule.
So, now it is possible to define several schedules and group
the datasets that are using the same schedule.

The configuration of this schedule-to-datasets mapping is
done in the `GOB_SCHEDULES` enviroment variable.

The DynamicDagRun operator (that triggers the next DAG) has been changed.
The previous version was only able to trigger the next dag in a
series of DAGs with a common prefix.

Now, the operator can be configured with a `labels_group`.
These labels are used make a subselection of the set of dags
with a common prefix.