Fix handling of cli parameters for Task runs.
Since Airflow 2.0, the cli parameters are provided in a different way.
This PR fixes the handling of parameters, so that cli task tests
can use parameters again.

This is especially useful for tweaking the `max_records` or the
`cursor_pos` parameters.
For cli task runs, the default number of records is 10000.
This can be changed with:

`airflow tasks test <dagname> <taskname> -t '{"max_records": 50000}'`