AB#27762 Fix Airflow TableCopyOperator to be able to handle view
In AB#23896 a new approach is implemented for the reference database.
This is one postgresql schema per dataset instead of all
datasets in the `public` schema. This will only be the case
for new datasets, existing dataset will stay in `public` for now.

However, the TableCopyOperator was breaking when views are being
introduced. The operator is used to copy data from a temporary
table to a target table. However, the assumptions is that
only regular tables are being used.

This PR solves this problem by checking for the target table
if it is a regular table or a view. If it turns out to be a view,
the underlying table (that the view references) is being used
as the target table.

This operator is doing all kinds of (too) clever things with
checking the set of column names and even the ordering.
Furthermore, it drops the target if it is not compatible
with the incoming source data (and structure of the columns of the
source data).

For now, dropping the target is raising an exception when the
target is a view. We do not want to add even more complexity
to this operator.

The operator does implicit schema migrations by dropping
the target table and replacing it with the source table.
This is an approach we want to get rid of anyway.

If related tables are associated with the source and target
of the copy operations, these targets are also checked.
If the targets are views, the same 'do-not-drop' rule applies.

There is one set of unittests in the Airflow repo handling
the TableCopyOperator. Additional unittests are added for
the new view-checking functionality.