[FIX] the DAG hoofdroutes_verkeer crashed due changed logic schema-tools
The DAG hoofdroutes_verkeer could not execute due some changes some of the functions in schema-tools. However, this was also an upportunity to simplify the DAG defintion and use more common operators and steps that where missing. Also renamed hoofdroutes_verkeer to hoofdroutes and overwrited the orginal hoofdroutes (since the latter was a legacy DAG anyway).