[NEW] workload wegenbestand Amsterdam
There has been a change in the source data with regards of the datasets related to the accesibility (and restrictions on it). This includes data set hoofdroutes. But also new tables are added.

Therefore a new DAG is created to ingest the new data source. The `hoofdroutes` DAG still remains open until also dataportaal can make the shift the the new endpoint.