SLACK MESSAGE for Azure
Since we are working on Azure with OTAP and also Airflow per solution, we need to know what DAG has been running where and for which team. Since we still have to deal with a hybrid situation with regards to CouldVPS (which has a different setup) and Azure, some hydrid modifications where set in place. When fully migrated, it can be adsjusted.

Also removed an old patch `MonkeyPatchedSlackWebhookHook` that was not needed any more.